+++ PACECloudSDK/AppKit/AppManagement/AppManager.swift
@@ -20,60 +20,29 @@ class AppManager {
     weak var delegate: AppManagerDelegate?
 
     private var locationProvider: AppDrawerLocationProvider
-    private var geoAPIManager: GeoAPIManager
     private var isLocationFetchRunning = false
     private var isGeneralFetchRunning = false
 
     private var currentlyDisplayedApps: (apps: [AppKit.AppData], location: CLLocation)?
 
-    private lazy var oneTimeLocationProvider: OneTimeLocationProvider = .init()
-
     init() {
         locationProvider = AppDrawerLocationProvider()
-        geoAPIManager = GeoAPIManager()
-
         locationProvider.delegate = self
     }
 
     func setConfigValues() {
         locationProvider.lowAccuracy = PACECloudSDK.shared.config?.allowedLowAccuracy ?? Constants.Configuration.defaultAllowedLowAccuracy
-        oneTimeLocationProvider.lowAccuracy = PACECloudSDK.shared.config?.allowedLowAccuracy ?? Constants.Configuration.defaultAllowedLowAccuracy
-        geoAPIManager.speedThreshold = PACECloudSDK.shared.config?.speedThreshold ?? Constants.Configuration.defaultSpeedThreshold
-        geoAPIManager.geoAppsScope = PACECloudSDK.shared.config?.geoAppsScope ?? Constants.Configuration.defaultGeoAppsScope
     }
 
     func startRetrievingLocation() {
         locationProvider.startFetchingLocation()
     }
-
-    func isPoiInRange(with id: String, completion: @escaping (Bool) -> Void) {
-        DispatchQueue.main.async { [weak self] in
-            self?.oneTimeLocationProvider.requestLocation { [weak self] location in
-                guard let location = location else {
-                    completion(false)
-                    return
-                }
-
-                self?.geoAPIManager.isPoiInRange(with: id, near: location, completion: completion)
-            }
-        }
-    }
-
-    func cofuGasStations(option: AppKit.CofuGasStation.Option, completion: @escaping ([AppKit.CofuGasStation]?) -> Void) {
-        geoAPIManager.cofuGasStations(option: option) { result in
-            guard case let .success(stations) = result else {
-                completion(nil)
-                return
-            }
-            completion(stations)
-        }
-    }
 }
 
 // MARK: - API requests
 extension AppManager {
-    private func cofuGasStations(for location: CLLocation, completion: @escaping ([AppKit.CofuGasStation]?) -> Void) {
-        geoAPIManager.locationBasedCofuStations(for: location) { [weak self] result in
+    private func cofuGasStations(for location: CLLocation, completion: @escaping ([AppKit.AppData]?) -> Void) {
+        POIKit.locationBasedCofuStations(for: location) { [weak self] result in
             switch result {
             case .failure(let error):
                 switch error {
@@ -85,12 +54,13 @@ extension AppManager {
                 }
 
             case .success(let stations):
-                completion(stations)
+                let appDatas = self?.retrieveAppData(from: stations)
+                completion(appDatas)
             }
         }
     }
 
-    private func fetchAppsByLocation(with location: CLLocation, completion: @escaping (([AppKit.CofuGasStation]?) -> Void)) {
+    private func fetchAppsByLocation(with location: CLLocation, completion: @escaping (([AppKit.AppData]?) -> Void)) {
         guard !isLocationFetchRunning else {
             completion(nil)
             delegate?.passErrorToClient(.fetchAlreadyRunning)
@@ -119,17 +89,22 @@ extension AppManager {
                     return
                 }
 
-                let cofuStations: [AppKit.CofuGasStation] = appsResponse.reduce(into: []) { result, app in
-                    guard let gasStationReferences = app.attributes?.references, let url = app.attributes?.pwaUrl else { return }
-
-                    let apps: [AppKit.CofuGasStation] = gasStationReferences.map { reference in
-                        .init(id: reference, coordinates: nil, polygon: nil, properties: ["apps": ["url": url]])
+                let appDatas: [AppKit.AppData] = appsResponse.reduce(into: []) { result, app in
+                    guard let id = app.id, let attributes = app.attributes, let gasStationReferences = attributes.references else {
+                        return
                     }
 
-                    result.append(contentsOf: apps)
+                    // In case if the identical app contains multiple gas station references
+                    let appDatas: [AppKit.AppData] = gasStationReferences.map {
+                        let metadata: [AppKit.AppMetadata: AnyHashable] = [AppKit.AppMetadata.appId: app.id, AppKit.AppMetadata.references: [$0]]
+                        let appData = AppKit.AppData(appID: id, appUrl: attributes.pwaUrl, metadata: metadata)
+                        return appData
+                    }
+
+                    result.append(contentsOf: appDatas)
                 }
 
-                completion(cofuStations)
+                completion(appDatas)
 
             case .failure(let error):
                 AppKitLogger.e("[AppManager] failed fetching local apps with error \(error)")
@@ -234,7 +209,7 @@ extension AppManager {
         URLBuilder.buildAppStartUrl(with: appUrl, decomposedParams: [.references], references: reference)
     }
 
-    private func retrieveAppData(for cofuGasStations: [AppKit.CofuGasStation], with location: CLLocation) {
+    private func retrieveAppData(from cofuGasStations: [POIKit.CofuGasStation]) -> [AppKit.AppData] {
         let appDatas: [AppKit.AppData] = cofuGasStations.reduce(into: []) { result, station in
             let metadata: [AppKit.AppMetadata: AnyHashable] = [AppKit.AppMetadata.references: [station.id]]
             let appDatas: [AppKit.AppData] = (station.properties["apps"] as? [[String: Any]] ?? []).map { app in
@@ -244,7 +219,7 @@ extension AppManager {
             result.append(contentsOf: appDatas)
         }
 
-        process(appDatas: appDatas, location: location)
+        return appDatas
     }
 
     private func process(appDatas: [AppKit.AppData], location: CLLocation) {
@@ -297,9 +272,9 @@ extension AppManager: AppDrawerLocationProviderDelegate {
     func didReceiveLocation(_ location: CLLocation) {
         AppKitLogger.i("[App Manager] Did receive location. Checking for available apps...")
 
-        cofuGasStations(for: location) { [weak self] apps in
-            guard let apps = apps else { return }
-            self?.retrieveAppData(for: apps, with: location)
+        cofuGasStations(for: location) { [weak self] appDatas in
+            guard let appDatas = appDatas else { return }
+            self?.process(appDatas: appDatas, location: location)
         }
     }